requirements.txt
instagrapi
python-dotenv
Pillow
Flask
waitress